Moved kind types to its own module
